refactor: Platform backend improvements and refinements
Backend service and middleware improvements:

**API & Middleware**:
- Enhanced app boundary middleware
- Improved API routing and middleware
- Better error handling and logging

**Core Services**:
- HTTP client robustness improvements
- Enhanced rate limiting service
- Better auth core and routing

**Service Enhancements**:
- Analytics service and aggregators improvements
- Audit router enhanced logging
- Ticketing service and workflow improvements
- CRM service updates

**Settings & Configuration**:
- Updated platform settings
- Cache models improvements
- Better contact schemas

**Testing**:
- Enhanced payment test fixtures
- Improved conftest for services

All changes maintain backward compatibility and improve reliability.

+++ src/dotmac/platform/cache/models.py
@@ -9,7 +9,7 @@ from enum import Enum
 from typing import Any
 from uuid import UUID, uuid4
 
-from sqlalchemy import Boolean, ForeignKey, Integer, String, Text
+from sqlalchemy import Boolean, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.dialects.postgresql import ENUM as SQLEnum
 from sqlalchemy.dialects.postgresql import JSON, TIMESTAMP
 from sqlalchemy.dialects.postgresql import UUID as PostgresUUID
@@ -159,11 +159,11 @@ class CacheStatistics(Base, TimestampMixin, TenantMixin):  # type: ignore[misc]
     total_requests: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     cache_hits: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     cache_misses: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    hit_rate: Mapped[float] = mapped_column(Integer, nullable=False, default=0)  # Percentage
+    hit_rate: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)  # Percentage (0.0-100.0)
 
     # Performance metrics
-    avg_hit_latency_ms: Mapped[float] = mapped_column(Integer, nullable=False, default=0)
-    avg_miss_latency_ms: Mapped[float] = mapped_column(Integer, nullable=False, default=0)
+    avg_hit_latency_ms: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
+    avg_miss_latency_ms: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
 
     # Storage metrics
     keys_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
